chore: convert source to CommonJs and publish source directly
Allow us to work on the code as it's going to run
instead of the hybrid ConnonJS-ESM format that cannot run anywhere
natively.
`babel-plugin-add-module-exports` always overwrote module.exports,
so the module.exports.default that Babel created was not
visible to consumers.
I've tested importing the code from CommonJS and native ESM.